TS for useQuery response
